refactor: moved csp configuration files into different namespace
