+++ src/UIPages/CspConfiguration/CspConfigurationEdit/CspConfigurationEditSection.cs
@@ -1,5 +1,5 @@
 ﻿using Kentico.Xperience.Admin.Base;
-using XperienceCommunity.CSP.UIPages;
+using XperienceCommunity.CSP.UIPages.CspConfiguration;
 
 [assembly: UIPage(
     parentType: typeof(CspConfigurationListing),
@@ -9,7 +9,7 @@ using XperienceCommunity.CSP.UIPages;
     templateName: TemplateNames.SECTION_LAYOUT,
     order: 300)]
 
-namespace XperienceCommunity.CSP.UIPages;
+namespace XperienceCommunity.CSP.UIPages.CspConfiguration;
 
 public class CspConfigurationEditSection : EditSectionPage<CSPConfigurationInfo>
 {
